requirements.txt
prometheus_client==0.20.0
requests==2.31.0
PyYAML==6.0.1